requirements.txt
apache-airflow[sendgrid]
apache-airflow-providers-docker
docker
